@@ -1,4 +1,4 @@
-<script>
+<script lang="ts">
 	import { Stack, Sidebar } from 'creditdesign-svelte-components';
 </script>
 
