@@ -1,4 +1,4 @@
-<script>
+<script lang="ts">
 	import Header from '$lib/Header/index.svelte';
 	import Footer from '$lib/Footer/index.svelte';
 </script>
